update: merge two scripts
