+++ package.json
@@ -62,9 +62,8 @@
         "build:categories": "ts-node src/js/cli.ts",
         "build:workflows": "node workflows/scripts/build_workflows.js",
         "build:applications": "node applications/scripts/build_applications.js",
-        "build:templates": "node applications/scripts/build_templates.js",
         "build:runtime-data": "node build_runtime_data.js",
-        "build:dist": "npm run transpile && npm run build:workflows && npm run build:applications && npm run build:templates && npm run build:runtime-data",
+        "build:dist": "npm run transpile && npm run build:workflows && npm run build:applications && npm run build:runtime-data",
         "lint": "eslint --cache src/js tests/js && prettier --write src/js tests/js",
         "lint:fix": "eslint --fix --cache src/js tests/js && prettier --write src/js tests/js",
         "lint:staged": "lint-staged",
